Refactor hook to function
